Major cleanups towards next release.
I cut almost all of the old cruft that we don't use anymore or
those bits we haven't implemented yet.

This means that reading the code (hopefully) makes sense
again.

I did some pedantic unset in 'for MOD in ...' to make it
behave more like a subshell. There are only two or three vars
that are allowed to 'get out' that for. (Mainly VIEW_LOG and
OUTPUT_TEXT).

Signed-off-by: Fernando J. Pereda <ferdy@gentoo.org>