Introduce xpatheval a XPath Swiss Army Knife.
Motivated by how broken xmllint --shell is we needed a nice
way to evaluate a number of XPath expressions on a number of
XML files.

That's exactly what xpatheval does, you give it an arbitrary
number of files and an arbitrary number of expressions and it
evaluates each expression on every file printing each result
in a new line. It is clever enough to guess how many files and
expressions you gave it.

[ an arbitrary number means at least 1 ]

Signed-off-by: Fernando J. Pereda <ferdy@gentoo.org>