links: Use xpatheval and fix some old bugs.
links.module was not properly handling anchor detection
because I was too lazy testing it in the first time. It
handles anchors properly and skips GuideXML thingies since it
cannot check them yet.

Also uses doxpath to guess if an anchor exists and to extract
a list of links in a document.

Signed-off-by: Fernando J. Pereda <ferdy@gentoo.org>